# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a GO REST server for sentix inference.",
        "title": "Swagger ginference-server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "chethanMysore",
            "url": "https://chethanmysore.github.io/portfolio/#contact",
            "email": "willishardrock94@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "performs user authentication and returns JWT Auth token on success",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register new user for inference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "Create User",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/models": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find all AI Models subscribed to the ginference-server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Get all models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AIModel"
                            }
                        }
                    }
                }
            }
        },
        "/models/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Subscribe new AIModel for inference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Create new model",
                "parameters": [
                    {
                        "description": "Create AIModel",
                        "name": "AIModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AIModelCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.AIModel"
                        }
                    }
                }
            }
        },
        "/models/edit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a subscribed AIModel's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Edit a model",
                "parameters": [
                    {
                        "description": "Update AIModel",
                        "name": "AIModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AIModelUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AIModel"
                        }
                    }
                }
            }
        },
        "/models/id/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find an AI Model using the given modelID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Search model by modelID",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AIModel"
                        }
                    }
                }
            }
        },
        "/models/name/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find AI Models matching the given model name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Search model by model name",
                "parameters": [
                    {
                        "maxLength": 18,
                        "minLength": 2,
                        "type": "string",
                        "description": "Model Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AIModel"
                            }
                        }
                    }
                }
            }
        },
        "/models/username/{username}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find AI Models created by the user with the given username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AIModels"
                ],
                "summary": "Search model by username",
                "parameters": [
                    {
                        "maxLength": 18,
                        "minLength": 5,
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AIModel"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AIModel": {
            "type": "object",
            "required": [
                "createdBy",
                "modelName"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "modelID": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                },
                "modifiedAt": {
                    "type": "string"
                }
            }
        },
        "model.AIModelCreate": {
            "type": "object",
            "required": [
                "createdBy",
                "modelName"
            ],
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                }
            }
        },
        "model.AIModelUpdate": {
            "type": "object",
            "required": [
                "modelID",
                "modelName"
            ],
            "properties": {
                "modelID": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                }
            }
        },
        "user.User": {
            "type": "object",
            "required": [
                "countryCode",
                "emailID",
                "firstName",
                "lastName",
                "phone",
                "userName"
            ],
            "properties": {
                "countryCode": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "emailID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                },
                "modifiedAt": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "userName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 5
                }
            }
        },
        "user.UserCreate": {
            "type": "object",
            "required": [
                "countryCode",
                "emailID",
                "firstName",
                "lastName",
                "password",
                "phone",
                "userName"
            ],
            "properties": {
                "countryCode": {
                    "type": "string"
                },
                "emailID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                },
                "userName": {
                    "type": "string",
                    "maxLength": 18,
                    "minLength": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}